pyproject.toml
[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.10.0"
openai = "^1.0"
speechrecognition = "^3.8.1"
pyttsx3 = "^2.90"
pytesseract = "^0.3.7"
pillow = "^9.1.1"
pymupdf = "^1.19.6"

[tool.poetry]
name = "multimodal-assistant"
version = "0.1.0"